feat(migrate): introduce SchemaEngineWasm; rename SchemaEngine to SchemaEngineCLI; make both classes comply to SchemaEngine interface; conditionally create SchemaEngine instances from Migrate, based on whether an "adapter" is given; adjust db/migrate commands and tests
