+++ packages/migrate/src/__tests__/rpc.test.ts
@@ -19,7 +19,7 @@ describe('applyMigrations', () => {
     ctx.fixture('existing-db-1-migration')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const migrationsList = await listMigrations(migrate.migrationsDirectoryPath!)
     const result = migrate.engine.applyMigrations({
       migrationsList,
@@ -37,7 +37,7 @@ describe('applyMigrations', () => {
     ctx.fixture('existing-db-brownfield')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const migrationsList = await listMigrations(migrate.migrationsDirectoryPath!)
     const result = migrate.engine.applyMigrations({
       migrationsList,
@@ -56,7 +56,7 @@ describe('applyMigrations', () => {
 describe('createDatabase', () => {
   it('should succeed - ConnectionString - sqlite', async () => {
     ctx.fixture('schema-only-sqlite')
-    const migrate = new Migrate()
+    const migrate = await Migrate.setup({})
     const result = migrate.engine.createDatabase({
       datasource: {
         tag: 'ConnectionString',
@@ -75,7 +75,7 @@ describe('createDatabase', () => {
   it('should succeed - Schema - postgresql', async () => {
     ctx.fixture('schema-only')
     const schemaContext = await loadSchemaContext()
-    const migrate = new Migrate()
+    const migrate = await Migrate.setup({})
     const result = migrate.engine.createDatabase({
       datasource: {
         tag: 'Schema',
@@ -98,7 +98,7 @@ describe('createMigration', () => {
     ctx.fixture('schema-only-sqlite')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = migrate.createMigration({
       migrationName: 'my_migration',
       draft: false,
@@ -117,7 +117,7 @@ describe('createMigration', () => {
     ctx.fixture('existing-db-1-migration')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = migrate.createMigration({
       migrationName: 'draft_123',
       draft: true,
@@ -136,7 +136,7 @@ describe('createMigration', () => {
 describe('dbExecute', () => {
   it('should succeed - sqlite', async () => {
     ctx.fixture('schema-only-sqlite')
-    const migrate = new Migrate()
+    const migrate = await Migrate.setup({})
     const result = migrate.engine.dbExecute({
       datasourceType: {
         tag: 'url',
@@ -157,7 +157,7 @@ describe('devDiagnostic', () => {
     ctx.fixture('schema-only-sqlite')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const migrationsList = await listMigrations(migrate.migrationsDirectoryPath!)
     const result = migrate.engine.devDiagnostic({
       migrationsList,
@@ -177,7 +177,7 @@ describe('devDiagnostic', () => {
     ctx.fixture('existing-db-1-migration-conflict')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const migrationsList = await listMigrations(migrate.migrationsDirectoryPath!)
     const result = migrate.engine.devDiagnostic({
       migrationsList,
@@ -212,7 +212,7 @@ describe('diagnoseMigrationHistory', () => {
     ctx.fixture('existing-db-1-migration')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const migrationsList = await listMigrations(migrate.migrationsDirectoryPath!)
     const result = migrate.engine.diagnoseMigrationHistory({
       migrationsList,
@@ -234,7 +234,7 @@ describe('diagnoseMigrationHistory', () => {
     ctx.fixture('existing-db-1-migration')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const migrationsList = await listMigrations(migrate.migrationsDirectoryPath!)
     const result = migrate.engine.diagnoseMigrationHistory({
       migrationsList,
@@ -257,7 +257,7 @@ describe('ensureConnectionValidity', () => {
   it('should succeed when database exists - SQLite', async () => {
     ctx.fixture('schema-only-sqlite')
     ctx.fs.write('dev.db', '')
-    const migrate = new Migrate()
+    const migrate = await Migrate.setup({})
     const result = migrate.engine.ensureConnectionValidity({
       datasource: {
         tag: 'ConnectionString',
@@ -273,7 +273,7 @@ describe('ensureConnectionValidity', () => {
     ctx.fixture('schema-only')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = migrate.engine.ensureConnectionValidity({
       datasource: {
         tag: 'Schema',
@@ -287,7 +287,7 @@ describe('ensureConnectionValidity', () => {
 
   it('should fail when database does not exist - SQLite', async () => {
     ctx.fixture('schema-only-sqlite')
-    const migrate = new Migrate()
+    const migrate = await Migrate.setup({})
     const result = migrate.engine.ensureConnectionValidity({
       datasource: {
         tag: 'ConnectionString',
@@ -306,7 +306,7 @@ describe('ensureConnectionValidity', () => {
 
   it('should fail when server does not exist - PostgreSQL', async () => {
     ctx.fixture('schema-only-sqlite')
-    const migrate = new Migrate()
+    const migrate = await Migrate.setup({})
     const result = migrate.engine.ensureConnectionValidity({
       datasource: {
         tag: 'ConnectionString',
@@ -334,7 +334,7 @@ describe('evaluateDataLoss', () => {
     ctx.fixture('schema-only-sqlite')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const migrationsList = await listMigrations(migrate.migrationsDirectoryPath!)
     const result = migrate.engine.evaluateDataLoss({
       migrationsList,
@@ -356,7 +356,7 @@ describe('evaluateDataLoss', () => {
     ctx.fixture('existing-db-1-migration')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const migrationsList = await listMigrations(migrate.migrationsDirectoryPath!)
     const result = migrate.engine.evaluateDataLoss({
       migrationsList,
@@ -379,7 +379,7 @@ describe('getDatabaseVersion - PostgreSQL', () => {
     ctx.fixture('schema-only')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = migrate.engine.getDatabaseVersion()
     await expect(result).resolves.toContain('PostgreSQL')
     migrate.stop()
@@ -388,7 +388,7 @@ describe('getDatabaseVersion - PostgreSQL', () => {
   it('[SchemaPath] should succeed', async () => {
     ctx.fixture('schema-only')
     const { schemas } = (await getSchemaWithPath())!
-    const migrate = new Migrate()
+    const migrate = await Migrate.setup({})
     const result = migrate.engine.getDatabaseVersion({
       datasource: {
         tag: 'Schema',
@@ -400,7 +400,7 @@ describe('getDatabaseVersion - PostgreSQL', () => {
   })
 
   it('[ConnectionString] should succeed', async () => {
-    const migrate = new Migrate()
+    const migrate = await Migrate.setup({})
     const result = migrate.engine.getDatabaseVersion({
       datasource: {
         tag: 'ConnectionString',
@@ -419,7 +419,7 @@ describe('markMigrationRolledBack', () => {
 
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
 
     const resultMarkRolledBacked = migrate.engine.markMigrationRolledBack({
       migrationName: '20201014154943_init',
@@ -439,7 +439,7 @@ describe('markMigrationRolledBack', () => {
     ctx.fixture('existing-db-1-migration')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = await migrate.createMigration({
       migrationName: 'draft_123',
       draft: true,
@@ -502,7 +502,7 @@ describe('markMigrationApplied', () => {
     ctx.fixture('existing-db-1-migration')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = await migrate.createMigration({
       migrationName: 'draft_123',
       draft: true,
@@ -534,7 +534,7 @@ describe('schemaPush', () => {
     ctx.fixture('schema-only-sqlite')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = migrate.engine.schemaPush({
       force: false,
       schema: toSchemasContainer(schemaContext.schemaFiles),
@@ -554,7 +554,7 @@ describe('schemaPush', () => {
     ctx.fixture('existing-db-1-draft')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = migrate.engine.schemaPush({
       force: false,
       schema: toSchemasContainer(schemaContext.schemaFiles),
@@ -574,7 +574,7 @@ describe('schemaPush', () => {
     ctx.fixture('existing-db-brownfield')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = migrate.engine.schemaPush({
       force: false,
       schema: toSchemasContainer(replaceInSchemas(schemaContext.schemaFiles, 'Blog', 'Something')),
@@ -596,7 +596,7 @@ describe('schemaPush', () => {
     ctx.fixture('existing-db-brownfield')
     const schemaContext = await loadSchemaContext()
     const { migrationsDirPath } = inferDirectoryConfig(schemaContext)
-    const migrate = new Migrate(schemaContext, migrationsDirPath)
+    const migrate = await Migrate.setup({ migrationsDirPath, schemaContext })
     const result = migrate.engine.schemaPush({
       force: true,
       schema: toSchemasContainer(replaceInSchemas(schemaContext.schemaFiles, 'Blog', 'Something')),
